+++ src/components/common/HomeComponents/homeBlog2/HomeBlog2.css
@@ -1,0 +1,190 @@
+/*Contenedor principal*/
+.HomeBlog-container2{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding-bottom: 5rem;
+    background-color: rgb(245,245,245);
+}
+
+/*Contenedor de título y descripción*/
+.HomeBlog-textContainer{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+/*Título*/
+.HomeBlog-title{
+    font-size: 2.5rem;
+    font-family: "Archivo";
+    color: var(--black);
+}
+
+/*Descripción*/
+.HomeBlog-description{
+    color: grey;
+    font-family: "InterTight";
+    text-align: center;
+    padding-top: 1.25rem;
+    width: 53.125rem;
+}
+
+/*Contenedor de bloques de noticias*/
+.HomeBlog-main-NoticiasContainer{
+    display: flex;
+    gap: 1.875rem;
+    justify-content: center;
+    margin-left: 5%;
+    margin-right: 5%;
+}
+
+/*Contenedor de noticias*/
+.HomeBlog-NoticiasContainer{
+    display: flex;
+    justify-content: center;
+    gap: 1.875rem;
+}
+
+/*Contenedor indivudual de imágen y título*/
+.HomeBlog-noticia{
+    position: relative;
+    width: 18.75rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+/*Imágen de noticia*/
+.HomeBlog-img{
+    width: 100%;
+    height: 25rem;
+    min-height: 200px;
+    object-fit: cover;
+    border-radius: 1.25rem;
+    margin-top: 4.375rem;
+}
+
+/*Título de noticia*/
+.HomeBlog-noticiaTitle{
+    font-family: "interTight";
+    color: var(--black);
+    padding-top: 1.25rem;
+}
+
+/*Link del título de la noticia*/
+a.link-blog{
+    transition: all 0.5s ease;
+}
+
+/*Contenedor del botón*/
+.HomeBlog-buttonContainer{
+    width: 14.375rem;
+    margin-top: 3.125rem;
+}
+
+/*Hovers*/
+.HomeBlog-noticia:hover a.link-blog{
+    transform: translatey(-100px);
+    color:#fff;
+    font-size: 28px;
+}
+.HomeBlog-noticia:hover a.link-blog h3{
+    color:#fff;
+}
+.HomeBlog-noticia:hover .HomeBlog-img{
+    filter: brightness(70%);
+}
+
+/*----------------Media Queries----------------*/
+
+@media(max-width:1350px){
+    .HomeBlog-noticia{
+        width: 15.625rem;
+    }
+    .HomeBlog-img{
+        height: 21.875rem;
+        margin-top: 3.125rem;
+    }
+}
+
+@media(max-width:1150px){
+    .HomeBlog-noticia{
+        width: 15rem;
+    }
+    .HomeBlog-img{
+        height: 21.25rem;
+        margin-top: 3.125rem;
+    }
+}
+
+@media(max-width:1110px){
+    .HomeBlog-main-NoticiasContainer{
+        flex-direction: column;
+        gap: 0%;
+    }
+    .HomeBlog-noticia{
+        width: 18.75rem;
+    }
+    .HomeBlog-img{
+        height: 25rem;
+    }
+    .HomeBlog-description{
+        width: 80%;
+    }
+}
+
+@media(max-width:760px){
+    .HomeBlog-title{
+        font-size: 2.1875rem;
+    }
+    .HomeBlog-description{
+        width: 80%;
+    }
+    .HomeBlog-NoticiasContainer{
+        flex-direction: column;
+        gap: 0%;
+    }
+    .HomeBlog-img{
+        height: 400px;
+    }
+    .HomeBlog-noticia{
+        width: 18.75rem;
+    }
+}
+
+@media(max-width:700px){
+    .HomeBlog-buttonContainer{
+        width: 80%;
+    }
+}
+
+@media(max-width:650px){
+    .HomeBlog-container2{
+        padding-bottom: 4rem;
+    }
+}
+
+@media(max-width:600px){
+    .HomeBlog-description{
+        width: 90%;
+    }
+}
+
+@media(max-width:450px){
+    .HomeBlog-img{
+        height: 21.875rem;
+    }
+    .HomeBlog-noticia{
+        width: 15.625rem;
+    }
+    .HomeBlog-buttonContainer{
+        width: 90%;
+    }
+}
+
+@media(max-width:400px){
+    .HomeBlog-title{
+        font-size: 1.875rem;
+    }
+}
