+++ src/components/common/HomeComponents/homeServices3/HomeServices3.css
@@ -1,0 +1,106 @@
+/*Contenedor Principal*/
+.homeServicios-container{
+    background-color: rgb(245,245,245);
+    padding-top: 5rem;
+    padding-bottom: 5rem;
+}
+
+/*Logo*/
+.homeSeervicios-logo{
+    width: 12.5rem;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+/*Título*/
+.homeServicios-titile{
+    text-align: center;
+    font-family: "archivo";
+    color: var(--black);
+}
+
+/*Texto*/
+.homeServicios-text{
+    margin-top: 1.25rem;
+    margin-bottom: 1.875rem;
+    font-family: "interTight";
+    color: grey;
+    width: 62.5rem;
+    margin-left: auto;
+    margin-right: auto;
+    text-align: center;
+    font-size: 1.125rem;
+}
+
+/*Enlace a NAVE*/
+.homeServicios-link{
+    text-align: center;
+    margin-top: 1.875rem;
+    font-family: "interTight";
+    font-size: 1.25rem;
+    color: var(--violet);
+}
+
+/*----------------Media Queries----------------*/
+
+@media(max-width:1250px){
+    .homeSeervicios-logo{
+        width: 11.25rem;
+    }
+}
+
+@media(max-width:1110px){
+    .homeServicios-link{
+        margin-top: 3.75rem;
+    }
+    .homeServicios-text{
+        width: 80%;
+    }
+}
+
+@media(max-width:950px){
+    .homeSeervicios-logo{
+        width: 10rem;
+    }
+    .homeServicios-titile{
+        font-size: 1.25rem;
+    }
+}
+
+@media(max-width:850px){
+    .homeServicios-link{
+        font-size: 1rem;
+    }
+}
+
+@media(max-width:700px){
+    .homeSeervicios-logo{
+        width: 8.75rem;
+    }
+    .homeServicios-titile{
+        font-size: 1.25rem;
+    }
+}
+
+@media(max-width:650px){
+    .homeServicios-container{
+        padding-top: 4rem;
+        padding-bottom: 4rem;
+    }
+    
+}
+
+@media(max-width:600px){
+    .homeServicios-text{
+        width: 90%;
+    }
+}
+
+@media(max-width:500px){
+    .homeServicios-link{
+        margin-top: 2.8125rem;
+    }
+    .homeServicios-text{
+        margin-bottom: .625rem;
+    }
+}
