+++ src/components/common/NaveComponents/Card/Card.css
@@ -1,73 +1,138 @@
 /*Card contenedor principal*/
 .card{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 28.125rem;
-    width: 18.75rem;
-    border-radius: 5px;
+    height: 23.125rem;
+    width: 21.875rem;
+    border-radius: 1.25rem;
+    background-color: var(--black);
     box-shadow: -2px 3px 5px rgba(0, 0, 0, 0.2), 2px 3px 5px rgba(0, 0, 0, 0.2);
 }
 
-/*Contenedor de nombre de servicio y precio*/
-.card-servicioContainer{
-    background-color: var(--violet);
+/*Contenedor de servicio*/
+.card-servicio-container{
+    background-color: var(--black);
+    border-top-left-radius: 1.25rem;
+    border-top-right-radius: 1.25rem;
+}
+
+/*Nombre del servicio*/
+.card-servicio{
+    color: white;
+    text-align: center;
+    font-family: "interTight";
+    font-weight: bold;
+    font-size: 1.5625rem;
+    padding: 1.875rem;
+}
+
+/*Contenedor de precio*/
+.card-info-container {
+    height: 17.5rem;
+    border-top-left-radius: 50% 20px;
+    border-top-right-radius: 50% 20px;
+    border-bottom-left-radius: 1.25rem;
+    border-bottom-right-radius: 1.25rem;
+    background: linear-gradient(to top, #f0ebff 0%, #fcfbff 95%, #ffffff 100%);
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    height: 9.375rem;
-    width: 100%;
-    gap: .625rem;
-    border-top-left-radius: 5px;  
-    border-top-right-radius: 5px;
 }
 
-/*Servicio*/
-.card-servicio{
-    color: white;
+/*Precio de lista*/
+.card-precio-promo{
     font-family: "interTight";
-    font-size: 1.25rem;
+    color: grey;
+    word-spacing: .3125rem
+}
+
+/*Precio tachado*/
+.tachado{
+    text-decoration: line-through;
 }
 
 /*Precio*/
 .card-precio{
-    color: white;
+    color: var(--violet);
     font-family: "interTight";
-    font-size: 1.875rem;
+    font-weight: bold;
+    font-size: 2.8125rem;
+    letter-spacing: .125rem;
+    margin-top: .625rem;
 }
 
-/*Contenedor de botones*/
-.card-buttonsContainer{
-    display: flex;
-    flex-direction: column;;
-    justify-content: space-evenly;
-    height: 9.375rem;
-    border-bottom: .0625rem solid var(--violet);
+/*Texto del precio*/
+.card-precio-text{
+    font-family: "interTight";
+    color: var(--violet);
+    font-size: 1.375rem;
+    letter-spacing: .0625rem;
 }
 
-/*Contenedor del logo*/
-.card-logoContainer{
-    display: flex;
-    justify-content: center;
-    align-items: end;
-    padding-bottom: 1.25rem;
-    height: 9.375rem;
-    border-bottom-right-radius: 5px;
-    border-bottom-left-radius: 5px;  
+/*Botón*/
+.card-button-container{
+    width: 15.625rem;
+    height: 3.4375rem;
+    margin-top: .9375rem;
+}
+
+/*Promo*/
+.card-promo{
+    font-family: "interTight";
+    font-weight: bold;
+    font-size: .875rem;
+    color: var(--darkgreen);
+    background-color: #a8e7c3;
+    padding-left: .625rem;
+    padding-right: .625rem;
+    padding-top: .3125rem;
+    padding-bottom: .3125rem;
+    border-radius: 1.25rem;
+    margin-top: .9375rem;
 }
 
 /*----------------Media Queries----------------*/
 
-@media(max-width:460px){
+@media(max-width:1250px){
     .card{
-        height: 18.75rem;
-        width:200px ;
+        transform: scale(0.9);
     }
-    .card-servicio{
-        font-size: .9375rem;
+}
+
+@media(max-width:1110px){
+    .card{
+        transform: scale(1);
     }
-    .card-precio{
-        font-size: 1.25rem;
+}
+
+@media(max-width:950px){
+    .card{
+        transform: scale(0.9);
+    }
+}
+
+@media(max-width:820px){
+    .card{
+        transform: scale(1);
+    }
+}
+
+@media(max-width:500px){
+    .card{
+        transform: scale(0.9);
+    }
+}
+
+@media(max-width:440px){
+    .card{
+        width: 18.75rem;
+    }
+    .card-button-container{
+        width: 12.5rem;
+    }
+}
+
+@media(max-width:385px){
+    .card{
+        width: 16.25rem;
     }
 }
