--- src/components/common/segundoHome/homeBlog2/HomeBlog2.css
@@ -1,132 +1,0 @@
-/*Contenedor principal*/
-.HomeBlog-container2{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding-top: 0rem;
-    padding-bottom: 5rem;
-    background-color: rgb(245,245,245);
-}
-
-/*Contenedor de título y descripción*/
-.HomeBlog-textContainer{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-/*Título*/
-.HomeBlog-title{
-    font-size: 2.5rem;
-    font-family: "Archivo";
-    color: var(--black);
-}
-
-/*Descripción*/
-.HomeBlog-description{
-    color: grey;
-    font-family: "InterTight";
-    text-align: center;
-    padding-top: 1.25rem;
-    margin-left: 20%;
-    margin-right: 20%;
-}
-
-/*Contenedor de noticias*/
-.HomeBlog-NoticiasContainer{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    margin-left: 5%;
-    margin-right: 5%;
-    gap: 1.875rem;
-    margin-bottom: 2.5rem;
-}
-
-/*Contenedor indivudual de imágen y título*/
-.HomeBlog-noticia{
-    width: 18.75rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-/*Imágen de noticia*/
-.HomeBlog-img{
-    width: 100%;
-    height: 25rem;
-    min-height: 200px;
-    object-fit: cover;
-    border-radius: 1.25rem;
-    margin-top: 4.375rem;
-}
-
-/*Título de noticia*/
-.HomeBlog-noticiaTitle{
-    font-family: "interTight";
-    color: var(--black);
-    padding-top: 1.25rem;
-}
-
-/*Contenedor del botón*/
-.HomeBlog-buttonContainer{
-    width: 14.375rem;
-}
-
-/*----------------Media Queries----------------*/
-
-@media(max-width:787px){
-    .HomeBlog-container{
-        padding-top: 4rem;
-        padding-bottom: 4rem;
-    }
-    .HomeBlog-description{
-        margin-left: 6%;
-        margin-right: 6%;
-    }
-    .HomeBlog-img{
-        width: 15.625rem;
-    }
-}
-
-@media(max-width:768px){
-    .HomeBlog-container{
-        align-items: normal;
-    }
-    .HomeBlog-buttonContainer{
-        width: 90%;
-        margin: auto;
-    }
-}
-
-@media(max-width:588px){
-    .HomeBlog-NoticiasContainer{
-        flex-direction: column;
-        gap: 0%;
-        margin-bottom: 3.125rem ;
-    }
-    .HomeBlog-noticia{
-        width: 100%;
-    }
-}
-
-.HomeBlog-noticia{
-    position: relative;
-}
-a.link-blog{
-    transition: all 0.5s ease;
-}
-.HomeBlog-noticia:hover a.link-blog{
-    transform: translatey(-100px);
-    color:#fff;
-    font-size: 28px;
-}
-.HomeBlog-noticia:hover a.link-blog h3{
-    color:#fff;
-}
-.HomeBlog-noticia:hover .HomeBlog-img{
-    filter: brightness(70%);
-
-}
-
-
