--- src/components/common/segundoHome/homeGuardia2/HomeGuardia2.css
@@ -1,103 +1,0 @@
-/*Contenedor principal*/
-.HomeGuardia-container2{
-    background-color: rgb(245,245,245);
-    padding-top: 0;
-    padding-bottom: 5rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.imagenGuardia2{
-    background-image: url(../../../../assets/images/pexels-chris-f-38966-10130754.jpg);
-    background-position: center;
-    background-size: cover;
-    background-repeat: no-repeat;
-    width: 37.5rem;
-    height: 500px;
-    border-radius: 1.25rem;
-}
-/*Contenedor de título y texto*/
-.HomeGuardia-infoContainer2{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    padding: 0 20px;
-    max-width: 50%;
-    margin:0;
-    width: 37.5rem;
-}
-
-/*Título*/
-.HomeGuardia-title2{
-    font-family: "Archivo";
-    font-size: 2.5rem;
-    color: #000;
-}
-
-/*Fondo de color "Eléctrica"*/
-.color-title2{
-    background:#8048ff;
-    color:#fff;
-    padding: 2px 15px;
-    clip-path: polygon(0 0, 95% 0, 100% 100%, 5% 100%);
-}
-
-/*Texto*/
-.HomeGuardia-description2{
-    margin-left: 0;
-    margin-right: 0;
-    font-family: "InterTight";
-    text-align: justify;
-    font-size: 1rem;
-    line-height: 1.5rem;
-    color: #000;
-}
-
-/*Contenedor de botón*/
-.HomeGuardia-buttonContainer2{
-    margin-top: 1.875rem;
-    width: 14.375rem;
-}
-
-/*----------------Media Queries----------------*/
-
-@media(max-width:768px){
-    .HomeGuardia-container2{
-        align-items: normal;
-    }
-    .HomeGuardia-buttonContainer2{
-        width: 90%;
-    }
-}
-
-@media(max-width:750px){
-    .HomeGuardia-container2{
-        padding-top: 4rem;
-        padding-bottom: 4rem;
-    }
-    .HomeGuardia-description2{
-        margin-left: 6%;
-        margin-right: 6%;
-    }
-}
-
-@media(max-width:600px){
-    .HomeGuardia-title2{
-        font-size: 2.1875rem;
-        text-align: center;
-    }
-    .HomeGuardia-description2{
-        font-size: .875rem;
-    }
-}
-
-@media(max-width:494px){
-    .HomeGuardia-title2{
-        font-size: 1.875rem;
-    }
-    .HomeGuardia-description2{
-        font-size: .875rem;
-    }
-}
