--- src/components/common/segundoHome/homePortada/HomePortada2.css
@@ -1,168 +1,0 @@
-/*Contenedor principal con imágen*/
-.background-container {
-    position: relative;
-    height: 100vh;
-  }
-  
-  .background-video {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    filter:brightness(50%); /* Ajusta la opacidad del video */
-    z-index: 1;
-  }
-  
-
-/*Oscuridad para imágen de fondo*/
-.homePortada-container {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.3);
-    z-index: 2;
-}
-
-
-/*Contenedor de contacto*/
-.homePortada-contactosContainer{
-    display: flex;
-    justify-content: space-evenly;
-}
-
-/*Contacto e ícono*/
-.homePortada-contactos{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: .3125rem;
-    text-decoration: none;
-}
-
-/*Texto de contacto*/
-.homePortada-contactosText{
-    color: white;
-    font-family: "InterTight";
-    margin-top: .9375rem;
-    margin-bottom: .9375rem;
-}
-
-/*Contenedor de textos*/
-.homePortada-bienvenidaContainer{
-display: flex;
-flex-direction: column;
-align-items: flex-start;
-margin-top: 11.25rem;
-padding-left: 13%;
-gap: .9375rem;
-}
-
-/*Título principal*/
-.homePortada-title{
-    color: white;
-    font-size: 3.75rem;
-    font-family: "archivo";
-}
-
-/*Descripción*/
-.homePortada-description{
-    color: white;
-    font-family: "InterTight";
-    width: 50%;
-}
-
-/*Contenedor del botón*/
-.homePortada-buttonContainer{
-    margin-top: 2.1875rem;
-    width: 12.5rem;
-    margin-left: 13%;
-}
-
-
-/*----------------Media Queries----------------*/
-
-@media(max-width:1386px){
-    .homePortada-title{
-        font-size: 3.125rem;
-    }
-}
-
-@media(max-width:1155px){
-    .homePortada-title{
-        font-size: 2.8125rem;
-    }
-}
-
-@media(max-width:1039px){
-    .background-container{
-        background-image: 
-        linear-gradient(rgba(22,22, 22, 0.0), rgba(22, 22, 22, 0.0)),
-        url("https://imageio.forbes.com/specials-images/imageserve/5e18ab57a854780006e8413a/0x0.jpg?format=jpg&height=900&width=1600&fit=bounds");
-    }
-}
-
-@media(max-width:867px){
-    .homePortada-contactosText{
-        font-size: .8125rem;
-    }
-}
-
-@media(max-width:805px){
-    .homePortada-description{
-        width: 90%;
-    }
-}
-
-@media(max-width:760px){
-    .homePortada-contactosContainer{
-        flex-direction: column;
-        gap: .625rem;
-        margin-bottom: .9375rem;
-        margin-top: .9375rem;
-    }
-    .homePortada-contactosText{
-        font-size: .8125rem;
-    }
-    .homePortada-title{
-        font-size: 2.1875rem;
-    }
-    .homePortada-contactosText{
-        margin: 0%;
-    }
-    #homePortada-tel{
-        display: none;
-    }
-}
-
-@media(max-width:768px){
-    .homePortada-bienvenidaContainer{
-        width: 90%;
-        margin-top: 7.5rem;
-    }
-    .homePortada-buttonContainer{
-        width: 60%;
-    }
-}
-@media(max-width:577px){
-    .homePortada-bienvenidaContainer{
-        width: 85%;
-    }
-}
-
-@media(max-width:480px){
-    .homePortada-title{
-        font-size: 1.875rem;
-    }
-}
-
-@media(max-width:409px){
-    .homePortada-title{
-        font-size: 1.5625rem;
-    }.homePortada-description{
-        font-size: .9375rem;
-    }
-}
