--- src/components/common/segundoHome/navbarHome/NavbarHome.css
@@ -1,28 +1,0 @@
-.logo-header{
-    width: 13.125rem;
-}
-
-/*----------------Media Queries----------------*/
-
-@media(max-width:1199px){
-    .logo-header{
-        width: 11.25rem;
-    }
-}
-
-
-@media(max-width:899px){
-    .logo-header{
-        width: 9.375rem;
-        padding-bottom: 1.25rem;
-    }
-}
-
-@media(max-width:599px){
-    .logo-header{
-        width: 7.5rem;
-        padding-bottom: 1.875rem;
-    }
-}
-
-
