+++ src/components/pages/FirstHome/FirstHome.css
@@ -1,0 +1,46 @@
+/* Estilos para el modal del Home */
+.modal-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100vw;
+    height: 100vh;
+    background: rgba(0, 0, 0, 0.7);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    z-index: 1000;
+    opacity: 1;
+}
+
+/* Contenido del modal */
+.modal-content {
+    background-color: var(--black);
+    padding: 10px;
+    border-radius: 20px;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.5);
+    max-width: 80%;
+    max-height: 80%;
+    overflow: hidden;
+    display: flex;
+    justify-content: center;
+    opacity: 0;
+    animation: fadeIn 0.5s forwards; 
+}
+
+/* Imagen del modal */
+.modal-image {
+    max-width: 100%;
+    height: auto;
+    border-radius: 30px;
+}
+
+/* Animación de desvanecimiento */
+@keyframes fadeIn {
+    from {
+        opacity: 0; 
+    }
+    to {
+        opacity: 1; 
+    }
+}
