+++ src/components/pages/FirstHome/homeGuardia/HomeGuardia.css
@@ -1,0 +1,100 @@
+/*Contenedor principal*/
+.HomeGuardia-container{
+    background-image: 
+    linear-gradient(rgba(22,22, 22, 0.7), rgba(22, 22, 22, 0.7)),
+    url("https://tec-eos.com/v3/wp-content/uploads/2023/03/personas-instalando-el-servicio-de-energia-electrica.webp"); 
+    background-size: cover;
+    background-repeat: repeat;
+    background-position: center 80%;
+    padding-top: 5rem;
+    padding-bottom: 5rem;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+/*Contenedor de título y texto*/
+.HomeGuardia-infoContainer{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 1.25rem;
+    max-width: 75rem;
+    margin-left: .625rem;
+    margin-right: .625rem;
+}
+
+/*Título*/
+.HomeGuardia-title{
+    font-family: "Archivo";
+    font-size: 2.5rem;
+    color: white;
+}
+
+/*Fondo de color "Eléctrica"*/
+.color-title{
+    background:#8048ff;
+    color:#fff;
+    padding: 2px 15px;
+    clip-path: polygon(0 0, 95% 0, 100% 100%, 5% 100%);
+}
+
+/*Texto*/
+.HomeGuardia-description{
+    margin-left: 13%;
+    margin-right: 13%;
+    font-family: "InterTight";
+    text-align: justify;
+    font-size: 1rem;
+    line-height: 1.5rem;
+    color: #fff;
+}
+
+/*Contenedor de botón*/
+.HomeGuardia-buttonContainer{
+    margin-top: 1.875rem;
+    width: 14.375rem;
+}
+
+/*----------------Media Queries----------------*/
+
+@media(max-width:768px){
+    .HomeGuardia-container{
+        align-items: normal;
+    }
+    .HomeGuardia-buttonContainer{
+        width: 90%;
+    }
+}
+
+@media(max-width:750px){
+    .HomeGuardia-container{
+        padding-top: 4rem;
+        padding-bottom: 4rem;
+    }
+    .HomeGuardia-description{
+        margin-left: 6%;
+        margin-right: 6%;
+    }
+}
+
+@media(max-width:600px){
+    .HomeGuardia-title{
+        font-size: 2.1875rem;
+        text-align: center;
+    }
+    .HomeGuardia-description{
+        font-size: .875rem;
+    }
+}
+
+@media(max-width:494px){
+    .HomeGuardia-title{
+        font-size: 1.875rem;
+    }
+    .HomeGuardia-description{
+        font-size: .875rem;
+    }
+}
