+++ src/components/pages/FirstHome/homeServicios/HomeServicios.css
@@ -1,0 +1,288 @@
+/* Contenedor principal */
+.homeServicios-container {
+    background: linear-gradient(to bottom left, #1a1a1a, #161616);
+    padding-top: 5rem;
+    padding-bottom: 5rem;
+    border-bottom: .0625rem solid var(--lightgreen);
+    border-top: .0625rem solid var(--lightgreen);
+    overflow: hidden;
+}
+
+/*Contenedor del logo*/
+.homeServicios-logo{
+    width: 9.375rem;
+    margin: 0 auto;
+}
+
+/*Título bajo el logo*/
+.homeServicios-title{
+    color: white;
+    font-family: "archivo";
+    text-align: center;
+    font-size: 1.25rem;
+}
+
+/* Contenedor de descripciones */
+.homeServicios-description {
+    margin-top: 2.5rem;
+    display: flex;
+    justify-content: center;
+    flex-wrap: wrap;
+    gap: 3.125rem;
+    align-items: center;
+}
+
+/* Descripción */
+.homeServicios-text {
+    color: var(--lightgreen);
+    border: .125rem solid var(--lightgreen);
+    font-family: "archivo";
+    font-size: 1.0625rem;
+    border-radius: .3125rem;
+    padding: 1.25rem;
+    width: 21.875rem;
+    text-align: center;
+}
+
+/* Contenedor de los botones */
+.homeServicios-buttonContainer {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 1.5625rem;
+    margin-top: 2.5rem;
+    justify-content: center;
+}
+
+/* Reset de botón */
+.button-services {
+    background: none;
+    cursor: pointer;
+    border: none;
+}
+
+/* Estilo del botón */
+.button-info-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background-color: white;
+    height: 11.25rem;
+    width: 11.25rem;
+    border: .125rem solid var(--lightgreen);
+    border-radius: .3125rem;
+    transition: background-color 0.3s, color 0.3s;
+}
+
+/* Texto dentro del botón */
+.button-text {
+    font-family: "intertight";
+    color: var(--black);
+    font-size: 2.5rem;
+    text-align: center;
+}
+
+/* Texto de Mbps */
+.megas {
+    color: grey;
+}
+
+/* Contenedor de información del servicio */
+.info-servicio-container {
+    padding: 1.875rem;
+    background-color:  white;
+    display: flex;
+    align-items: center;
+    gap: 3.125rem;
+    border-radius: 2.3125rem;
+    margin-bottom: 3.125rem;
+    margin-top: 1.875rem;
+    width: 63%;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+/* Contenedor del texto y el botón con precio*/
+.info-servicio-text-button{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 2.5rem;
+}
+
+/*Contenedor del nombre y descrip del servicio*/
+.info-servicio-text{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: .625rem;
+}
+
+/* Título del servicio */
+.info-servicio-title {
+    font-family: "interTight";
+    color: var(--darkgreen);
+    font-size: 1.5625rem;
+}
+
+/* Descripción del servicio */
+.info-servicio-description {
+    color: var(--black);
+    font-family: "interTight";
+    font-size: 1.125rem;
+    text-align: center;
+}
+
+/*Contenedor de precio y botón*/
+.contratar{
+    display: flex;
+    gap: 1.875rem;
+    align-items: center;
+    background-color: #8048ff;
+    padding: 1.875rem;
+    border-radius: .3125rem;
+    border:1px solid var(--darkgreen);
+}
+
+/*Precio del servicio*/
+.price{
+    font-family: "intertight";
+    color: #ffff;
+    font-size: 1.5625rem;
+}
+
+/* Imagen del servicio */
+.info-servicio-img {
+    max-width: 25rem; 
+    border-radius: .3125rem;
+}
+
+/* Efecto hover en los botones */
+.button-services:hover .button-info-container {
+    background-color: var(--lightgreen);
+    color: var(--white);
+    border-color: #8048ff
+}
+
+/*Hover de megas e icono */
+.button-services:hover .button-info-container .MuiSvgIcon-root,
+.button-services:hover .button-info-container .megas {
+    color: var(--black);
+}
+
+
+/* Contenedor de whatsapp */
+.whatsapp-container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: .3125rem;
+    text-decoration: none;
+    cursor: pointer;
+}
+
+/* Whatsapp */
+.whatsapp {
+    color: var(--lightgreen);
+    font-family: "intertight";
+    font-size: 1.5625rem;
+}
+/*----------------Media Queries----------------*/
+
+@media(max-width:1435px){
+    .button-info-container{
+        height: 9.375rem;
+        width: 9.375rem;
+    }
+    .button-text{
+        font-size: 2.1875rem;
+    }
+}
+
+@media(max-width:1220px){
+    .button-info-container{
+        height: 6rem;
+        width: 6rem;
+    }
+    .button-text{
+        font-size: 1.375rem;
+    }
+    .homeServicios-logo{
+        width: 8.125rem;
+    }
+    .homeServicios-title{
+        font-size: 1.0625rem;
+    }
+}
+
+@media(max-width:1100px){
+    .info-servicio-container{
+        flex-direction: column;
+        gap: 1.25rem;
+    }
+    .info-servicio-text-button{
+        gap: 1.25rem;
+    }
+}
+
+@media(max-width:908px){
+    .homeServicios-buttonContainer{
+        gap: .9375rem;
+    }
+}
+
+@media(max-width:820px){
+    .homeServicios-description{
+        flex-direction: column;
+        gap: 1.25rem;
+    }
+    .homeServicios-text{
+        width: 90%
+    }
+    .info-servicio-container{
+        width: 90%;
+    }
+}
+
+@media(max-width:750px){
+    .homeServicios-container{
+        padding: 4rem 0.5rem;
+    }
+    .contratar{
+        flex-direction: column;
+        gap:.625rem
+    }
+    .info-servicio-img{display: none;}
+    .whatsapp{
+        font-size: 1.25rem;
+    }
+    .homeServicios-logo{
+        width: 7.5rem;
+    }
+    .homeServicios-title{
+        font-size: .9375rem;
+    }
+}
+
+@media(max-width:450px){
+    .button-info-container{
+        height: 5rem;
+        width: 5rem;
+    }.button-text{
+        font-size: 1.125rem;
+    }.info-servicio-title{
+        font-size: 1.375rem;
+    }
+    .info-servicio-img{
+        display: none;
+    }
+    .homeServicios-logo{
+        width: 6.25rem;
+    }
+    .homeServicios-title{
+        font-size: .8125rem;
+    }
+    .homeServicios-text{
+        font-size: 1rem;
+    }
+}
