+++ src/components/pages/FirstHome/homeServicios2/HomeServicios2.css
@@ -1,0 +1,282 @@
+/* Contenedor principal */
+.homeServicios-container2 {
+    background: rgb(245, 245, 245);
+    padding-top: 5rem;
+    padding-bottom: 5rem;
+    border-color:transparent
+}
+
+/*Contenedor del logo*/
+.homeServicios-logo2{
+    width: 9.375rem;
+    margin: 0 auto;
+}
+.logonave2{
+    filter:invert(1)
+}
+
+/*Título bajo el logo*/
+.homeServicios-title2{
+    color: #000;
+    font-family: "archivo";
+    text-align: center;
+    font-size: 1.25rem;
+}
+
+/* Contenedor de descripciones */
+.homeServicios-description2 {
+    margin-top: 2.5rem;
+    display: flex;
+    justify-content: center;
+    flex-wrap: wrap;
+    gap: 3.125rem;
+    align-items: center;
+}
+
+/* Descripción */
+.homeServicios-text2 {
+    color: var(--lightgreen);
+    border: .125rem solid var(--lightgreen);
+    font-family: "archivo";
+    font-size: 1.0625rem;
+    border-radius: .3125rem;
+    padding: 1.25rem;
+    width: 21.875rem;
+    text-align: center;
+}
+
+/* Contenedor de los botones */
+.homeServicios-buttonContainer2 {
+    display: flex;
+    flex-wrap: wrap;
+    border-radius: 10px;
+    justify-content: center;
+    border:2px solid transparent;
+    border-radius: 10px;
+    margin-top:0px;
+    box-shadow: 3px 3px 3px rgba(0,0,0,0.2);
+}
+
+/* Reset de botón */
+.button-services2 {
+    background: none;
+    cursor: pointer;
+    border: none;
+}
+
+/* Estilo del botón */
+.button-info-container2 {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background-color: white;
+    height: 7em;
+    padding: 10px;
+    width: 11.25rem;
+    border-radius: .3125rem;
+    transition: background-color 0.3s, color 0.3s;
+}
+
+/* Texto dentro del botón */
+.button-text2 {
+    font-family: "intertight";
+    color: var(--black);
+    font-size: 1.5rem;
+    text-align: center;
+}
+
+/* Texto de Mbps */
+.megas {
+    color: grey;
+}
+
+/* Contenedor de información del servicio */
+.info-servicio-container2 {
+    padding: 1.875rem;
+    background-color:  white;
+    display: flex;
+    align-items: center;
+    gap: 3.125rem;
+    border-radius: 2.3125rem;
+    margin-bottom: 3.125rem;
+    margin-top: 1.875rem;
+    width: 63%;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+/* Contenedor del texto y el botón con precio*/
+.info-servicio-text-button2{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 2.5rem;
+}
+
+/*Contenedor del nombre y descrip del servicio*/
+.info-servicio-text2{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: .625rem;
+}
+
+/* Título del servicio */
+.info-servicio-title2 {
+    font-family: "interTight";
+    color: var(--darkgreen);
+    font-size: 1.5625rem;
+}
+
+/* Descripción del servicio */
+.info-servicio-description2 {
+    color: var(--black);
+    font-family: "interTight";
+    font-size: 1.125rem;
+    text-align: center;
+}
+
+/*Contenedor de precio y botón*/
+.contratar2{
+    display: flex;
+    gap: 1.875rem;
+    align-items: center;
+    background-color: transparent;
+    padding: 1.875rem;
+    border-radius: .3125rem;
+    border:1px solid transparent;
+}
+.contratar2 button{
+    border:2px solid var(--darkgreen);
+    border-radius: 10px;
+}
+/*Precio del servicio*/
+.price2{
+    font-family: "intertight";
+    color: #ffff;
+    font-size: 1.5625rem;
+}
+
+/* Imagen del servicio */
+.info-servicio-img2 {
+    max-width: 25rem; 
+    border-radius: .3125rem;
+}
+
+/* Efecto hover en los botones */
+.button-services2:hover .button-info-container2 {
+    background-color: #8048ff;
+    color: var(--white);
+    border-color: #8048ff
+}
+.button-services2 .button-info-container2{
+    border-color:#fff;
+}
+.button-services2:hover .button-text2{
+    color:#fff;
+}
+
+/*Hover de megas e icono */
+.button-services2:hover .button-info-container2 .MuiSvgIcon-root,
+.button-services2:hover .button-info-container2 .megas{
+    color: var(--white);
+}
+
+
+/* Contenedor de whatsapp */
+.whatsapp-container2 {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: .3125rem;
+    text-decoration: none;
+    cursor: pointer;
+}
+
+/* Whatsapp */
+.whatsapp2 {
+    color: var(--lightgreen);
+    font-family: "intertight";
+    font-size: 1.5625rem;
+}
+
+
+/*agregados*/
+.containerglobal2{
+    display: flex;
+    flex-direction: row-reverse;
+    margin: 20px auto;
+    max-width: 1200px;
+    padding: 20px;
+    gap: 10px;
+    
+}
+.homeServicios-buttonContainer2{
+    flex-direction: column;
+    align-items: center;
+    gap:1px;
+    background-color: #fff;
+    padding: 10px 20px;
+}
+.button-services2 {
+    width: 250px;
+}
+.button-services2 div{
+    width: 100%;
+}
+.tipoServicio2{
+    display: flex;
+    background-color: #fff;
+    padding: 10px 20px;
+    border-radius: 10px;
+    box-shadow: 3px 3px 3px rgba(0,0,0,0.2);
+}
+
+.info-servicio-container2.dinamico{
+    width: 300px;
+    border:2px solid #8048ff;
+    padding: 0;
+    overflow: hidden;
+    
+}
+.info-servicio-container2.dinamico .info-servicio-text2{
+    width: 100%;
+    background-color: var(--black);
+    color:#fff;
+}
+.info-servicio-container2.dinamico .info-servicio-text2 p{
+    color:#fff;
+    padding: 10px 10px 30px 10px;
+}
+.info-servicio-container2.dinamico .info-servicio-title2{
+    width: 100%;
+    height: 60px;
+    line-height: 60px;
+    background-color:#8048ff;
+    color:#fff;
+    text-align: center;
+    
+  
+}
+.info-servicio-container2.dinamico div.info-servicio-text-button2{
+    width: 100%;
+    
+}
+.info-servicio-container2.dinamico .contratar2{
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    
+}
+.info-servicio-container2.dinamico .contratar2 h5{
+    color:#000;
+}
+
+/*----------------Media Queries----------------*/
+
+@media (max-width: 1250px) {
+    .containerglobal2 {
+        transform: scale(0.9);
+    }
+}
