+++ src/components/pages/nave/Nave.jsx
@@ -15,6 +15,16 @@ const Nave = () => {
     dispatch(setTitle(''));
   }, [dispatch]);
 
+  const serviciosNave = [
+    { servicio: "Internet 50 MB", precio: "$12.614" },
+    { servicio: "Internet 100 MB", precio: "$16.846" },
+    { servicio: "Internet 300 MB", precio: "$19.864" },
+    { servicio: "Internet 500 MB", precio: "$24.893" },
+    { servicio: "Full TV", precio: "$8.570" },
+    { servicio: "Fútbol Premium", precio: "$8.786" },
+];
+
+
   return (
     <section className='nave-main-container'>
       <Fade triggerOnce={true} duration={800} delay={300}>
@@ -22,7 +32,7 @@ const Nave = () => {
         <p className='nave-text'>Descubre nuestras increíbles opciones de Internet y televisión diseñadas para satisfacer todas tus necesidades de entretenimiento y conectividad.</p>
       </Fade>
       <Fade triggerOnce={true} duration={800} delay={300} direction='up'>
-        <CardCarousel/>
+        <CardCarousel servicios={serviciosNave}  showArrowsAndDots={true}/>
       </Fade>
     </section>
   );
